ROSALIND-30 Add commit message formatter to pre-push
This has a few unpleasant properties (namely, we have to run git
commit --amend to modify the commit message, which means we have
to repush, and fail in pre-push).

An alternative is to simply always have 'git push' fail, and provide
an alternative submission mechanism.

Fixes https://github.com/cwgreene/rosalind/issues/30