ROSALIND-33 Don't count loading libraries against solutions
The solution to the fib algorithm (not fibd; that's a horrible hack
using sympy) is unfairly penalized for pulling in the numpy library.
I've added the imports numpy and sympy into the rosalind library; so
they'll be loaded once during setup before the tests are run.

We should in future come up with a clean way to measure the overhead of
pulling in a library, and import them all at once.

Fixes https://github.com/cwgreene/rosalind/issues/33